Merged PR 21072: Bug: Enable pipefail in pipeline
# 🔍 Description

This PR enables pipefail in the tasks/commands that need it.

Also removes template that is no longer being used.

# 🤔 Rationale

Normally the exit status is determined by the last command. Enabling pipefail returns a failure status if any command within it fails, not just the last one.

# 📝 Checks

- Example where task didn't fail before: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=676129&view=logs&j=f7e50092-5d43-5e2f-c0b6-53093257bb47&t=cd65b91d-460d-50fa-5ca5-1f0adb02059c&l=376

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug by enabling the `pipefail` option in various pipeline scripts to ensure that the pipeline fails if any command in a pipeline fails.
- `/.pipelines/templates/stages/build_docker_image/trident-container.yml`: Added `pipefail` to the `set` command.
- `/.pipelines/templates/stages/testing_baremetal/baremetal-testing.yml`: Added `pipefail` to the `set` command and a specific script.
- `/.pipelines/templates/stages/testing_common/trident-rebuild.yml`: Added `pipefail` to the `set` command.
- `/.pipelines/templates/stages/testing_servicing/vm-testing.yml`: Added `pipefail` to multiple `set` commands and specific scripts.
- `/.pipelines/templates/stages/testing_vm/netlaunch-testing.yml` and `netlaunch-prep.yml`: Added `pipefail` to the `set` command.

Related work items: #9857