Merged PR 21091: bug: stop SELinux from relabeling if not enabled
# 🔍 Description

This PR adds checks to ensure SELinux is in the OS and SELinux is enabled before running `setfiles` to relabel. Currently, if Trident is run on an image without SELinux we still try to run `setfiles`, which does not exist resulting in an error.

There are two cases in which SELinux should not run:

* Image does not have SELinux (`setfiles` not on OS)
* Image does have SELinux, but HC sets SELinux mode to `disabled`

In addition, we should return an `InvalidInput` error if there is a mismatch between the HC and the deployed image, i.e. HC enables SELinux but it is not on the image.

This PR also alphabetizes `InvalidInput` errors (error messages not edited).

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses an issue where the SELinux module attempts to relabel files regardless of the SELinux configuration status, causing failures on systems without SELinux.
- `src/subsystems/selinux.rs`: Added checks to ensure SELinux is enabled and `setfiles` exists before attempting to relabel files.
- `trident_api/src/error.rs`: Introduced a new error `SelinuxEnabledButNotFound` and reorganized existing error definitions for better clarity.

Related work items: #8212