Merged PR 21006: engineering: Trident SSH config to only include specific file
# 🔍 Description

Currently Trident's SSH config has an include directive to include all files in `/etc/ssh/sshd_config.d`. Only one file is in this directory, so changing the directive to include the specific name of the file.

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
Refines SSH configuration to include only a specific file for Trident instead of all files in the directory.
- `src/subsystems/osconfig/users.rs`: Replaces wildcard import with a specific file import in SSH configuration.
- `src/subsystems/osconfig/users.rs`: Introduces `GLOBAL_CONFIG_FILE_NAME` constant for the global user configuration file.

Related work items: #9087