Merged PR 21121: Bug: Install needed version of libtpms0
# 🔍 Description

Add version of libtpms0 needed for swtpm in ppa:stefanberger.

# 🤔 Rationale

Seems that there was an update in the version of swtpm in ppa:stefanberger which now requires the version 0.10.0 of libtpms0. This version is only available after adding the PPA.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug related to the installation of the required version of `libtpms0` to resolve a failure in turning on the machine during the netlaunch process.
- Updated `/.pipelines/templates/stages/testing_vm/netlaunch-prep.yml` to purge `libtpms0` and ensure its installation.

Related work items: #9911