Merged PR 21078: engineering: Log raid sync status if we are waiting for RAID to sync
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

It's helpful to log the sync status of the raid arrays if we are waiting to sync, that way the user can have a rough idea how much they might want to increase their `syncTimeout` field by.

Looks like this:
![image.png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/21078/attachments/image.png)

# 🤔 Rationale

Why is this PR needed?

Provides more actionable information for user when they wait for RAID syncing.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Log raid sync status if we are waiting to sync

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds logging of the RAID sync status when waiting for RAID arrays to sync, particularly on timeout.
- `src/engine/storage/raid.rs`: Added code to log the current RAID sync status by reading from `/proc/mdstat` and outputting the result to the debug log.

Related work items: #8744