Merged PR 21147: engineering: Download latest successful run-id only and don't allow partial success for container test
# 🔍 Description

This PR changes the Makefile to only query the latest successful run of the trident-ci pipeline in order to download images (reverting changes from !20672). Also, this PR makes it so that the container test doesn't allow for partial success.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request ensures that only the latest successful pipeline run ID is downloaded and disallows partial success for container tests.
- `Makefile`: Updated to download only the latest successful run ID and removed the allowance for partial success.
- `.pipelines/templates/stages/testing_vm/netlaunch-testing.yml`: Removed `continueOnError` for container runtime environment.

Related work items: #9767