Merged PR 21042: Engineering: Test azl 3.0 preview of container image with Trident
# 🔍 Description

We now can also test the preview in AzureLinux 3.0 in containers. The version of the container can be selected from pipeline run that build it: https://dev.azure.com/mariner-org/mariner/_build?definitionId=2116&_a=summary&tagFilter=3.0-preview (default is latest successful run).
The container is used to:
- Build Trident rpms
- Test trident running from the container

# 🤔 Rationale

We want to test azl changes with Trident. This PR adds the changes needed to test Trident running from the azl 3.0 preview of the container image.

# 📝 Checks

- Example of a run using preview: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=675183&view=results

# 📌 Follow-ups

- #9965
- #9966

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Enable testing with preview base image in CI and pre-release pipelines.

#### PR Summary
This pull request adds the capability to use a preview base image in CI and pre-release pipelines, allowing tests to be run on the preview image.
- Added `/.pipelines/templates/stages/common_tasks/preview-container.yml` to handle downloading and setting up the preview container image.
- Updated `/.pipelines/templates/stages/trident_rpms/trident-stage.yml` and `/.pipelines/templates/stages/build_docker_image/trident-container.yml` to include parameters for selecting base image type and build ID.
- Modified `/.pipelines/trident-ci.yml` and `/.pipelines/trident-prerelease.yml` to integrate the new preview base image functionality.
- Enhanced `/.pipelines/templates/stages/trident_rpms/build-source.yml` and `/.pipelines/templates/stages/trident_rpms/release.yml` to support the new parameters for preview base image.

Related work items: #9667