Merged PR 21032: engineering: add simple mirrored host config to e2e configs
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Added a mirrored host configuration for testing purposes. . This new setup uses RAID1 for the EFI System Partition (ESP), root, and trident partitions to provide redundancy, fault tolerance, and continuous operation in the event of disk failures. Currently, host configurations lack this mirrored setup, so this addition is crucial for testing these scenarios effectively.

1. **Redundancy for ESP**:
   - The ESP contains the boot loader and other files necessary for booting the operating system. By mirroring the ESP across multiple drives, you ensure that if one drive fails, the system can still boot from the other drive. This redundancy is crucial for maintaining the ability to start the system in case of a hardware failure.

2. **Redundancy for Root Partition**:
   - The root partition contains the core operating system and essential files. Mirroring this partition ensures that all data is available on an alternative drive if one drive fails. This helps maintain the integrity and availability of the operating system and prevents downtime.

3. **Redundancy for Trident Partition**:
For Trident to function properly, it requires access to the /var/lib/trident directory, where we store the datastore. To ensure Trident remains operational in the event of a disk failure, the Trident directory must be placed on a RAID1 array.

# 🤔 Rationale

Why is this PR needed?

Currently, the host configurations do not have this mirrored setup; hence, a new configuration is being added to test this case.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:[]()

- #0000

# 🗒️ Notes

add simple mirrored hc to e2e configs

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds a new mirrored host configuration to the end-to-end (e2e) test configurations.
- Added `trident-config.yaml` for mirrored host configuration in `e2e_tests/trident_configurations/mirrored/`
- Updated `target-configurations.yaml` to include the new mirrored configuration for both bare metal and virtual machine targets
- Added `test-selection.yaml` for mirrored configuration in `e2e_tests/trident_configurations/mirrored/`

Related work items: #9844