Merged PR 21174: Do not create machine-id on update
# 🔍 Description

Do not create `/etc/machine-id` on update, as it will get later overwritten by a copy from the current instance.

Also add a script and a HC sample to validate A/B update on verity setup.
Also add a makefile target to regenerate the verity images with the latest trident bits to simplify verity testing.

# 🤔 Rationale

Reduces unnecessary steps for verity setups and simplifies local validation.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to prevent the creation of machine-id during updates.

#### PR Summary
This pull request modifies the behavior to ensure that the machine-id is only created during a clean install and not during updates.
- `src/engine/storage/mod.rs`: Added a condition to check if the servicing type is a clean install before creating the machine-id.
- `Makefile`: Corrected a minor typo in a comment.