Merged PR 21241: engineering: Add testing base config in pr-e2e runs
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

Noticed we should be testing base instead of simple in the pr-e2e to have the base and combined pair tested on a PR. Will also be required for alerts to catch PR regressions.

# 🤔 Rationale

Why is this PR needed?

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Switch to testing base config instead of simple in pr-e2e runs

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request updates the end-to-end testing configuration to use the 'base' configuration instead of 'simple' for pull request runs.
- `e2e_tests/target-configurations.yaml`: Replaces 'simple' with 'base' in the pull request testing configurations.

Related work items: #10080