Merged PR 21312: bug: Fix two bugs in UKI update
Use create_dir_all to create the loader directory in the ESP so it doesn't fail if the directory already exists. And start indexes at 100 so that the default UKI name (`vmlinuz-6.6.29...`) isn't ordered after it. (The default isn't parsed as a Trident UKI entry, so its index isn't considered when selecting what index to use)

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses two bugs in the UKI update process.
- `src/engine/boot/esp.rs`: Fixes directory creation by using `fs::create_dir_all` instead of `fs::create_dir`.
- `src/engine/boot/esp.rs`: Sets the initial `max_index` value to 99 instead of 0.

Related work items: #6583