Merged PR 21272: engineering: enforce COSI filesystem types are known
# 🔍 Description

Enforce that `fs_type` field in COSI (as well as mock OS image) must be of type `OsImageFileSystemType`, in order to ensure that only filesystem types supported by Trident are accepted. This error will be caught when deserializing the metadata file. For validation of the OS image, we check using function to match (Host Configuration) FileSystemType to OsImageFileSystemType.

----
#### AI description  (iteration 1)
#### PR Classification
Code enhancement

#### PR Summary
This pull request enforces that only known filesystem types are accepted in COSI metadata to ensure compatibility with Trident's supported filesystems.
- Added `FromStr` and `Deserialize` implementations for `FileSystemType` in `trident_api/src/config/host/storage/filesystem.rs`.
- Updated `FileSystemImage` and related methods in `src/osimage/mod.rs` to use `FileSystemType` instead of `String`.
- Modified `validate_host_config` function in `src/engine/storage/osimage.rs` to handle `FileSystemType` directly.
- Changed `fs_type` field from `String` to `FileSystemType` in `src/osimage/cosi/metadata.rs` and `src/osimage/mock.rs`.

Related work items: #10082