Merged PR 21350: engineering: Use separate metric files for A/B update runs, and kill the netlisten process...
# 🔍 Description

What is this PR about? feature/doc/engineering/bug?

There is a transient issue where sometimes the netlisten port doesn't free up right away since it is ran in the background. This causes a metrics file issue with A/B update sometimes, so this PR makes sure we kill the netlisten process after an update is done, as well as use separately named files for update metrics to avoid transient file issues.

# 🤔 Rationale

Why is this PR needed?

Fix occasional port and file issues for A/B update runs in the pipelines.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

Use separate metric files for A/B update runs, and kill the netlisten process after each update finishes to ensure the port is free for next run

----
#### AI description  (iteration 1)
#### PR Classification
Code enhancement and bug fix.

#### PR Summary
This pull request introduces separate metric files for A/B update runs and ensures the `netlisten` process is terminated to prevent port clashes in subsequent stages.
- `.pipelines/templates/stages/testing_common/e2e-ab-update-stage-finalize-test-run.yml`: Added commands to kill the `netlisten` process to avoid port clashes.
- `.pipelines/templates/stages/testing_common/e2e-test-run.yml`: Updated metric file paths for A/B update runs and added commands to kill the `netlisten` process.
- `.pipelines/templates/stages/testing_common/trident-metrics.yml`: Modified conditions for metrics upload steps to remove the `succeeded()` check.

Related work items: #10225