Merged PR 21447: engineering: speed up baremetal tests
# 🔍 Description

baremetal tests are using an exponential backoff to test when machines come back online after updates.  it looks like there is a 4 minute sleep that we are just waiting through pretty consistently because of this.  removing the exponential backoff.  also reducing the script total timeout from (1+2+4+...+2^15) seconds to 10 minutes.

validated with trident-testing: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=698722&view=results

may not be apple-to-apple, but there does seem to be savings on a lot of the tests:
* trident-testing successful baremetal tests time: 8hr10min
* last few trident-prerelease successful baremetal tests times: 9hr54min, 9hr56min

# 🤔 Rationale

Some tests are waiting 3x for a machine to come back online.  Watching the console, it looks like a majority of the time, the machine comes back just as we start the 256 second wait.  Potentially, we could save close to 12 minutes on any of those tests.

Also, i suspect having an exponential wait time that gets to 2^15 seconds is incompatible with any test timeout we have.

# 📝 Checks

- [ ] Check [dev-docs/manual-validation.md](/dev-docs/manual-validation.md)

# 📌 Follow-ups

TODO:

- #0000

# 🗒️ Notes

make check-ssh-connection more responsive to speed up tests

----
#### AI description  (iteration 1)
#### PR Classification
Code optimization to speed up baremetal tests.

#### PR Summary
This pull request optimizes the SSH connection retry logic to speed up baremetal tests.
- `.pipelines/templates/stages/testing_common/scripts/check-ssh-connection.sh`: Changed the retry wait time to a fixed 5 seconds and set a total wait time of 10 minutes, removing the exponential backoff logic.

Related work items: #10355