Merged PR 21503: bug: Update sizes in trident-setup.yaml for FTs so that A/B volumes are even
# 🔍 Description

This PR updates the sizes in trident-setup.yaml for FTs so that A/B volumes are even.

Currently, our trident-setup.yaml that we use for the FTs does not pass the storage graph v2 validation:

[ERROR trident_api::config::host::storage] EXPERIMENTAL GRAPHv2: Failed to build storage graph: Referrer 'root' of kind 'ab-volume' references partitions of different sizes

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request updates the volume sizes in the `trident-setup.yaml` file to ensure that A/B volumes are even.
- `functional_tests/trident-setup.yaml`: Changed `root-a` size from 5G to 4G and `root-b` size from 2G to 4G.

Related work items: #10402