Merged PR 21492: bug: Allow for a null `source` in findmnt output
# 🔍 Description

This PR offers a fix to a failure where Trident would fail to parse the output of `findmnt` if the value of `source` is null. This fix is necessary because in the case that a mount point is an overlay, the `source` might be reported not as `overlay` but as `null` by the tool.

This is relevant to the scenario for LinuxGuard/steamboat where containerd and tardev-snapshotter are used to run Trident inside a container, such as here: https://dev.azure.com/mariner-org/mariner/_build/results?buildId=703548&view=logs&j=d8a50e79-ffc7-500b-0d95-4ec046c034d1&t=0314d650-6a8a-57b4-36c2-aa990003c5ff&s=330995f7-875f-5bff-99c0-87893ffbb8d9

TODO: Waiting for a confirmation from Jirka

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug by allowing the `source` field in `findmnt` output to be null.
- `osutils/src/findmnt.rs`: Changed the `source` field in `MountpointMetadata` to be an `Option<PathBuf>` to handle cases where a mounted filesystem does not have a specific source.
- Updated functional tests to handle the optional `source` field correctly.

Related work items: #10398