Merged PR 21547: feature: Enable A/B update testing for memory-constraint-combined HC test
# 🔍 Description

This PR enables the A/B update testing scenario for memory-constraint-combined HC test, in addition to the clean install test.

----
#### AI description  (iteration 2)
#### PR Classification
New feature to enable A/B update testing for memory-constraint-combined HC test.

#### PR Summary
This pull request enables A/B update testing for the memory-constraint-combined configuration in the HC test.
- Modified `/.pipelines/templates/stages/testing_common/e2e-test-run.yml` to always check for A/B update testing requirements.
- Updated `e2e_tests/helpers/ab_update.py` to handle additional error messages related to memory limits during A/B update testing.

Related work items: #10425