Merged PR 21590: bug: Allow log-explorer script to parse corrupted logs
# 🔍 Description

Currently if logs are corrupted such that there are extraneous new lines breaking up a log line, then the log-explorer script fails. This PR adds a warning message to the user and removes extraneous new lines as a means to fix the corrupted log file.

This script can be tested with the sample corrupted log file in the bug's description: #10113

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug in the log-explorer script that caused it to fail when parsing corrupted logs.
- `scripts/log-explorer`: Added logic to detect and attempt to recover from corrupted log files by removing unexpected newlines and issuing a warning.

Related work items: #10113