Merged PR 21686: Engineering: Update BM template
# 🔍 Description

Updates the stage dependencies of BareMetal template for container scenario.

# 🤔 Rationale

We need the dependencies change to be able to use the same template for host and container scenarios. (Change was already on VM template, Baremetal wasn't updated).

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request updates the BM template to improve code readability and organization.
- `.pipelines/templates/e2e-template.yml`: Adjusted indentation for better readability.
- `.pipelines/templates/stages/testing_baremetal/baremetal-testing.yml`: Reorganized job steps, including adding and reordering test image build steps.

Related work items: #10518