Merged PR 21704: engineering: Remove old target to download OS modifier
# 🔍 Description

This PR removes the outdated target to download OS Modifier in Makefile. This target is no longer applicable as it's not published to the feed anymore, so the dev/user is now expected to build the OS modifier via make build-osmodifier.

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request removes an outdated target for downloading the OS modifier binary from the Makefile.
- `Makefile`: Removed the target for downloading the `osmodifier` binary from the Azure DevOps artifacts feed.

Related work items: #10668