Merged PR 21736: engineering: Test whether Trident can re-run A/B update successfully with the same HC after a rollback
# 🔍 Description

This is a follow up PR to improve the E2E rollback test logic. In particular, after the rollback is validated, the test will stage and finalize another A/B update with the same Host Configuration, to make sure that the update succeeds on a re-run.

----
#### AI description  (iteration 1)
#### PR Classification
Test enhancement

#### PR Summary
This pull request enhances the test script to verify if Trident can re-run an A/B update successfully with the same HC after a rollback.
- `scripts/loop-update/loop-update.sh`: Added a command to display the update configuration before running the Trident stage operation.

Related work items: #10167