Merged PR 21705: engineering: Carry over hostname during A/B update
# 🔍 Description

During AB update, carry over the hostname from volume A to volume B if no hostname is specified in the host configuration. This functionality can be disabled with the internal parameter `disableHostnameCarryOver`.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a feature to carry over the existing machine hostname during an AB update if no hostname is provided in the Host Configuration.
- `src/subsystems/osconfig/mod.rs`: Added logic to carry over the hostname and updated the `OsConfigSubsystem` to store the previous hostname.
- `trident_api/src/error.rs`: Added a new error type `ReadHostname` for handling hostname read failures.
- `trident_api/src/constants.rs`: Introduced a new internal parameter `DISABLE_HOSTNAME_CARRY_OVER` to allow disabling the hostname carry-over functionality.

Related work items: #10550