Merged PR 21835: Bug: Fix Trident runtime environment
# 🔍 Description

This enables the verity and verity-raid test in trident-pr-e2e pipeline and fixes the selection for Trident Runtime environment.

# 🤔 Rationale

Recent changes tried using the variable `verityRequired` directly instead of defining it again, but the conditional in the tasks that defines the variable will cause an error for the cases that don't require A/B update.

Related work items: #10816