Merged PR 21844: feature: Support OS images in the logic that determines if A/B update is required
# 🔍 Description

This PR modifies the logic in the `ab_update_required()` (previously `needs_ab_update()`) to support A/B updates with OS images. For now, Trident will first compare if any partition images from the old API have been changed and then, if there is no change there, it will compare the OS images in the old and new specs.

TODO: As a follow up, we need to introduce hashes for the OS images. This is very important for security, as well as for determining if an A/B update is required in a more granular way. Related ADO task: https://dev.azure.com/mariner-org/ECF/_workitems/edit/10845

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds support for COSI in the logic to initiate A/B updates.
- `src/engine/storage/image/mod.rs`: Refactored `needs_ab_update` to `ab_update_required` and added logic to check OS images if COSI is enabled.
- `src/engine/storage/mod.rs`: Updated to use the new `ab_update_required` function.
- `trident_api/src/config/host/error.rs`: Added new error type `DeployPartitionImagesAfterOsImage` and updated error messages.
- `src/engine/storage/osimage.rs` and `trident_api/src/error.rs`: Updated error messages to use consistent naming for OS image related errors.

Related work items: #10821